containers/etl/requirements.txt
pandas
google-cloud-storage
fsspec
gcsfs
